\name{u.rlmerMod}
\alias{b}
\alias{b.lmerMod}
\alias{b.rlmerMod}
\alias{u}
\alias{u.lmerMod}
\alias{u.rlmerMod}
\title{Random effects accessor}
\usage{
  \method{u}{rlmerMod} (object, ...)

  \method{u}{lmerMod} (object, ...)

  \method{b}{rlmerMod} (object, ...)

  \method{b}{lmerMod} (object, ...)

  b(object, ...)

  u(object, ...)
}
\arguments{
  \item{object}{merMod object}

  \item{...}{ignored}
}
\description{
  Simple accessors for the (spherical) random effects.
}

