\name{len}
\alias{len}
\title{length}
\usage{
  len(x, what)
}
\arguments{
  \item{x}{merMod object}

  \item{what}{length is requested}
}
\description{
  Get various numbers of parameters / lengths of vectors.
}
\details{
  b, u: length of random effects vector beta, coef: number
  of fixed effects theta: length of vector theta r, e:
  number of observations
}

