% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFile.R
\name{loadAndMergePartialResults}
\alias{loadAndMergePartialResults}
\title{Load And Merge Partial Results}
\usage{
loadAndMergePartialResults(files)
}
\arguments{
\item{files}{vector of filenames (including paths) of files containing the
processed results}
}
\description{
Convenience function that loads the results stored in each of the files and
then calls \code{\link{mergeProcessedFits}} to merge them.
}
\seealso{
\code{\link{processDatasetsInParallel}}
}
\author{
Manuel Koller
}
