\name{nobs-methods}
\docType{methods}
\alias{nobs-methods}
\alias{nobs,lmerMod-method}
\title{ Number of observations }
\description{
  Returns the number of observations of the fitted data set.
}
\section{Methods}{
\describe{
\item{\code{signature(object = "lmerMod")}}{
  Returns the number of observations of the fitted data set.
}
}}
\keyword{methods}
