\name{other}
\alias{show}
\alias{show,Rcpp_SmoothPsi-method}
\alias{show,Rcpp_HuberPsi-method}
\alias{show,Rcpp_PsiFunction-method}
\alias{show,Rcpp_PsiFunction to Prop II PsiFunction wrapper-method}
\title{Other methods}
\usage{
\S4method{show}{Rcpp_SmoothPsi}(object)
\S4method{show}{Rcpp_HuberPsi}(object)
\S4method{show}{Rcpp_PsiFunction}(object)
\S4method{show}{`Rcpp_PsiFunction to Prop II PsiFunction wrapper`}(object)
}
\description{
  Other miscellaneous utilities for instances of the PsiFunction class.
}
\arguments{
  \item{object}{instance of class \code{PsiFunction} to be plotted}
}
\examples{
show(smoothPsi)
}
\keyword{utilities}