% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptoticEfficiency.R
\name{partialMoment_standardNormal}
\alias{partialMoment_standardNormal}
\title{Compute Partial Moments}
\usage{
partialMoment_standardNormal(z, n)
}
\arguments{
\item{z}{partial moment boundary, the expectation is taken from -Inf to z.}

\item{n}{which moment to compute, needs to be >= 2.}
}
\description{
Computes a partial moment for the standard normal distribution. This is the
expectation taken not from -Infinity to Infinity but just to \code{z}.
}
\examples{
  partialMoment_standardNormal(0, 2)
}
\references{
Winkler, R. L., Roodman, G. M., & Britney, R. R. (1972). The
Determination of Partial Moments. Management Science, 19(3), 290–296.
http://www.jstor.org/stable/2629511, equation (2.5)
}
