% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageDatasets.R
\name{saveDatasets}
\alias{saveDatasets}
\title{Save datasets}
\usage{
saveDatasets(datasets, path = getwd(), file, chunkSize)
}
\arguments{
\item{datasets}{dataset list generated by one of the generate functions.}

\item{path}{path to save the datasets to.}

\item{file}{filename to use, without extension.}

\item{chunkSize}{if provided, datasets are split into \code{chunkSize}
chunks and then saved.}
}
\value{
filename or vector of filenames.
}
\description{
Saves dataset to one or more files.
}
\details{
The file will be saved to \code{path/filename.Rdata}.

If \code{chunkSize} is not missing, the filename is interpreted as format
specifier and passed onto \code{\link[base]{sprintf}}. One argument is given, the
index of the chunk.
}
\author{
Manuel Koller
}
