\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,rlmerMod-method}
\title{ Show rlmerMod object}
\description{
  Print a summary information about rlmerMod object.
}
\section{Methods}{
\describe{
\item{\code{signature(object = "rlmerMod")}}{
  Print a summary information about rlmerMod object.
}
}}
\keyword{methods}
