% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiFunc2.R
\name{chgDefaults}
\alias{chgDefaults}
\alias{chgDefaults,psi_func_rcpp-method}
\title{Change default arguments}
\usage{
\S4method{chgDefaults}{psi_func_rcpp}(object, ...)
}
\arguments{
\item{object}{instance to convert}

\item{...}{arguments to change}
}
\description{
Change the default arguments for a psi_func_rcpp object
}
\note{
Note that names of named arguments are ignored. Only the order of the arguments
considered when assigning new arguments.
}
\examples{
sPsi <- chgDefaults(smoothPsi, k=2)
curve(sPsi@psi(x), 0, 3)
curve(smoothPsi@psi(x), 0, 3, col="blue", add=TRUE)
}
\keyword{utilities}
