% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiFunc2.R
\name{psi2propII}
\alias{psi2propII}
\alias{psi2propII,Rcpp_SmoothPsi}
\alias{psi2propII,psi_func_rcpp-method}
\title{Convert to Propsal II weight function}
\usage{
psi2propII(object, ...)

\S4method{psi2propII}{psi_func_rcpp}(object, ...)
}
\arguments{
\item{object}{instance of Rcpp_PsiFunction class to convert}

\item{...}{optional, new default arguments passed to chgDefaults.}
}
\description{
Converts the psi_func object into a function that corresponds
to Proposal II, i.e., a function of the squared weights.
The other elements of the psi_func object are adapted accordingly.
}
\examples{
par(mfrow=c(2,1))
plot(smoothPsi)
plot(psi2propII(smoothPsi))
}
\keyword{utilities}
