\name{psi2propII}
\alias{psi2propII}
\alias{psi2propII,psi_func-method}
\title{Convert to Propsal II weight function}
\arguments{
  \item{object}{psi_func object to convert}

  \item{...}{optional, new default arguments passed to
  chgDefaults.}
}
\description{
  Converts the psi_func object into a function that
  corresponds to Proposal II, i.e., a function of the
  squared weights. The other elements of the psi_func
  object are adapted accordingly.
}

