\docType{class}
\name{rlmerMod-class}
\alias{coef.rlmerMod}
\alias{fitted.rlmerMod}
\alias{formula.rlmerMod}
\alias{model.frame.rlmerMod}
\alias{model.matrix.rlmerMod}
\alias{print.rlmerMod}
\alias{print.summary.rlmer}
\alias{rlmerMod-class}
\alias{show.rlmerMod}
\alias{show,rlmerMod-method}
\alias{show.summary.rlmer}
\alias{summary.rlmerMod}
\alias{summary.summary.rlmer}
\alias{terms.rlmerMod}
\alias{update.rlmerMod}
\alias{vcov.rlmerMod}
\alias{vcov.summary.rlmer}
\title{rlmerMod Class
Class "rlmerMod" of Robustly Fitted Mixed-Effect Models

A robust mixed-effects model as returned by \code{\link{rlmer}}.}
\description{
  rlmerMod Class Class "rlmerMod" of Robustly Fitted
  Mixed-Effect Models

  A robust mixed-effects model as returned by
  \code{\link{rlmer}}.

  ##
}
\section{Objects from the Class}{
  Objects are created by calls to \code{\link{rlmer}}.
}
\examples{
showClass("rlmerMod")
}
\seealso{
  \code{\link{rlmer}}
}
\keyword{classes}

